+++ src/schick/seg102.cpp
@@ -387,7 +387,7 @@ void mspell_verwandlung(void)
 			g_spelltarget_e->flags.petrified = 0;
 
 			/* prepare message */
-			sprintf(g_dtp2,	get_tx(114), (uint8_t*)GUI_names_grammar((signed int)0x8000, g_spelltarget_e->mon_id, 1));
+			sprintf(g_dtp2,	get_tx(114), (uint8_t*)GUI_names_grammar(0x8000, g_spelltarget_e->mon_id, 1));
 		}
 
 	} else if (g_spelltarget_e->flags.mushroom) {
@@ -541,7 +541,7 @@ void mspell_blitz(void)
 		g_spelltarget_e->blind = 3;
 
 		/* prepare message */
-		sprintf(g_dtp2,	get_tx(85), (uint8_t*)GUI_names_grammar((signed int)0x8000, g_spelltarget_e->mon_id, 1));
+		sprintf(g_dtp2,	get_tx(85), (uint8_t*)GUI_names_grammar(0x8000, g_spelltarget_e->mon_id, 1));
 	}
 }
 
@@ -569,7 +569,7 @@ void mspell_eisenrost(void)
 				get_spelltarget()->inventory[HERO_INVENTORY_SLOT_RIGHT_HAND].flags.broken = 1;
 
 				/* prepare message */
-				sprintf(g_dtp2,	get_tx(92), (uint8_t*)GUI_names_grammar((signed int)0x8000, item_id, 0), get_spelltarget()->alias);
+				sprintf(g_dtp2,	get_tx(92), (uint8_t*)GUI_names_grammar(0x8000, item_id, 0), get_spelltarget()->alias);
 
 			} else {
 				g_monster_spell_ae_cost = -2;
@@ -589,7 +589,7 @@ void mspell_eisenrost(void)
 			g_spelltarget_e->weapon_broken = 1;
 
 			/* prepare message */
-			sprintf(g_dtp2,	get_tx(91), (uint8_t*)GUI_names_grammar((signed int)0x8000, g_spelltarget_e->mon_id, 1));
+			sprintf(g_dtp2,	get_tx(91), (uint8_t*)GUI_names_grammar(0x8000, g_spelltarget_e->mon_id, 1));
 		}
 	}
 }
@@ -726,7 +726,7 @@ void mspell_plumbumbarum(void)
 		g_spelltarget_e->at = g_spelltarget_e->at - 3;
 
 		/* prepare message */
-		sprintf(g_dtp2,	get_tx(95), (uint8_t*)GUI_names_grammar((signed int)0x8001, g_spelltarget_e->mon_id, 1));
+		sprintf(g_dtp2,	get_tx(95), (uint8_t*)GUI_names_grammar(0x8001, g_spelltarget_e->mon_id, 1));
 	}
 }
 
@@ -779,7 +779,7 @@ void mspell_paralue(void)
 		g_spelltarget_e->flags.petrified = 1;
 
 		/* prepare message */
-		sprintf(g_dtp2,	get_tx(103), (uint8_t*)GUI_names_grammar((signed int)0x8000, g_spelltarget_e->mon_id, 1));
+		sprintf(g_dtp2,	get_tx(103), (uint8_t*)GUI_names_grammar(0x8000, g_spelltarget_e->mon_id, 1));
 	} else {
 		/* target is a hero */
